backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
google-api-python-client = "^2.100.0"
transformers = "^4.34.0"
torch = "^2.1.0"
python-dotenv = "^1.1.1"
yt-dlp = "^2025.9.26"
torchaudio = "^2.8.0"
openai-whisper = "^20250625"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest-mock (>=3.15.1,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
